Detect if player has reached the exit door
This only triggers if the player overlaps while on the ground.

There's a bug with phaser right now with static bodies.
If a sprite overlaps with a static body, the onFloor flag is set to true.
That means that the win condition is triggered even in the air.
Not exactly what we wanted, so we had to change the door to have a dynamic body.