Create Prealoder scene
This scene will be run at the beginning to load all the assets.
There aren't much assets right now but the levels should add a few.
Once the fonts are loaded we move to the game scene.
If the wait time gets longer we'll add a loader to it.

The key constants are in their own file as well. Easier to manage