Player is stuck to moving platform
The player now follows a moving platform!
By keeping track of the platform we're on and adding their deltas, we stay on them.

We reset the player's collision flags every update.
This ensures we don't need fancy logic to remove them.
The caveat is that we need to calculate those flags every update.
Technically that happens anyway, and it's minor to not affect performance too much.